refactor: split up functionality across more files
